app/requirements.txt
numpy==1.22.4
pandas==1.5.3
pillow==8.4.0
protobuf==3.20.1
keras==2.12.0